+++ public/app/percona/check/components/AllChecksTab/AllChecksTab.tsx
@@ -10,8 +10,8 @@ import { CheckDetails, Interval } from 'app/percona/check/types';
 import { ExtendedColumn, FilterFieldTypes, Table } from 'app/percona/integrated-alerting/components/Table';
 import { AlertLocalStorage } from 'app/percona/shared/components/Elements/AlertLocalStorage/AlertLocalStorage';
 import { CustomCollapsableSection } from 'app/percona/shared/components/Elements/CustomCollapsableSection/CustomCollapsableSection';
+import { UpgradePlanWrapper } from 'app/percona/shared/components/Elements/CustomCollapsableSection/UpgradePlanWrapper/UpgradePlanWrapper';
 import { FeatureLoader } from 'app/percona/shared/components/Elements/FeatureLoader';
-import { UpgradePlanWrapper } from 'app/percona/shared/components/Elements/UpgradePlanWrapper/UpgradePlanWrapper';
 import { useCancelToken } from 'app/percona/shared/components/hooks/cancelToken.hook';
 import { usePerconaNavModel } from 'app/percona/shared/components/hooks/perconaNavModel';
 import { getPerconaSettingFlag } from 'app/percona/shared/core/selectors';
